requirements.txt
redis
fastapi
uvicorn
requests
python-dotenv
psycopg2-binary
pandas
numpy
python-dotenv
